IO-744 Project information to talpa form (#431)
Add form to enter information needed for opening sap project number. Form is intended to get the necessary information
to generate an excel to be sent to Talpa. Form is placed in a tab under project view.

After user has entered the necessary information and saves the form, excel is generated in the backend
and downloaded automatically.